Removing product service
